// workspace/projects/sys/src-com/app/tuninghelper/tuninghelper-analysis/tuninghelper-task-detail/domain/service-type.ts
/*
 * Copyright 2022 Huawei Technologies Co., Ltd
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

/** 业务类型 */
export enum ServiceTypeValue {
  Network = 'network',
  Cpu = 'cpu',
  Disk = 'disk'
}
export type ServiceType = {
  value: ServiceTypeValue;
  checked: boolean;
  label: string;
  disabled: boolean;
  notAllowed?: boolean;
};